orchestrator-agent/pyproject.toml
[project]
name = "octopets-orchestrator-agent"
version = "1.0.0"
description = "Multi-agent orchestrator for Octopets platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agent-framework-azure-ai>=0.1.0",
    "agent-framework-core>=0.1.0",
    "azure-identity>=1.19.0",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-sdk>=1.29.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = []
